* csort
* cmp functions, return s8
* sort functions, take a 'void const* user_data'

+++ source/main/cpp/c_qsort.cpp
@@ -53,26 +53,26 @@ namespace ncore
             }
         }
 
-        static inline u8* sMed3(u8* a, u8* b, u8* c, s32 (*cmp)(const void* const, const void* const, void*), void* data)
+        static inline u8* sMed3(u8* a, u8* b, u8* c, s8 (*cmp)(const void* const, const void* const, void*), void* data)
         {
             return cmp(a, b, data) < 0 ? (cmp(b, c, data) < 0 ? b : (cmp(a, c, data) < 0 ? c : a)) : (cmp(b, c, data) > 0 ? b : (cmp(a, c, data) < 0 ? a : c));
         }
-        static inline u16* sMed3(u16* a, u16* b, u16* c, s32 (*cmp)(const void* const, const void* const, void*), void* data)
+        static inline u16* sMed3(u16* a, u16* b, u16* c, s8 (*cmp)(const void* const, const void* const, void*), void* data)
         {
             return cmp(a, b, data) < 0 ? (cmp(b, c, data) < 0 ? b : (cmp(a, c, data) < 0 ? c : a)) : (cmp(b, c, data) > 0 ? b : (cmp(a, c, data) < 0 ? a : c));
         }
-        static inline u32* sMed3(u32* a, u32* b, u32* c, s32 (*cmp)(const void* const, const void* const, void*), void* data)
+        static inline u32* sMed3(u32* a, u32* b, u32* c, s8 (*cmp)(const void* const, const void* const, void*), void* data)
         {
             return cmp(a, b, data) < 0 ? (cmp(b, c, data) < 0 ? b : (cmp(a, c, data) < 0 ? c : a)) : (cmp(b, c, data) > 0 ? b : (cmp(a, c, data) < 0 ? a : c));
         }
-        static inline u64* sMed3(u64* a, u64* b, u64* c, s32 (*cmp)(const void* const, const void* const, void*), void* data)
+        static inline u64* sMed3(u64* a, u64* b, u64* c, s8 (*cmp)(const void* const, const void* const, void*), void* data)
         {
             return cmp(a, b, data) < 0 ? (cmp(b, c, data) < 0 ? b : (cmp(a, c, data) < 0 ? c : a)) : (cmp(b, c, data) > 0 ? b : (cmp(a, c, data) < 0 ? a : c));
         }
     }  // namespace __qsort
 
 	// Sort routine for element size other than 2, 4, or 8 bytes
-    void g_qsortN(u8* a, s32 n, s32 es, s32 (*cmp)(const void*, const void*, void*), void* user_data)
+    void g_qsortN(u8* a, s32 n, s32 es, s8 (*cmp)(const void*, const void*, void*), void* user_data)
     {
         u8 *pa, *pb, *pc, *pd, *pl, *pm, *pn;
         s32 d, r, swaptype, swap_cnt;
@@ -163,7 +163,7 @@ namespace ncore
         }
     }
 
-    void g_qsort2(u16* a, s32 n, s32 (*cmp)(const void*, const void*, void*), void* data)
+    void g_qsort2(u16* a, s32 n, s8 (*cmp)(const void*, const void*, void*), void* data)
     {
         u16 *pa, *pb, *pc, *pd, *pl, *pm, *pn;
         s32  d, r, swaptype, swap_cnt;
@@ -252,7 +252,7 @@ namespace ncore
         }
     }
 
-    void g_qsort4(u32* a, s32 n, s32 (*cmp)(const void*, const void*, void*), void* data)
+    void g_qsort4(u32* a, s32 n, s8 (*cmp)(const void*, const void*, void*), void* data)
     {
         u32 *pa, *pb, *pc, *pd, *pl, *pm, *pn;
         s32  d, r, swaptype, swap_cnt;
@@ -341,7 +341,7 @@ namespace ncore
         }
     }
 
-    void g_qsort8(u64* a, s32 n, s32 (*cmp)(const void*, const void*, void*), void* data)
+    void g_qsort8(u64* a, s32 n, s8 (*cmp)(const void*, const void*, void*), void* data)
     {
         u64 *pa, *pb, *pc, *pd, *pl, *pm, *pn;
         s32  d, r, swaptype, swap_cnt;
@@ -430,7 +430,7 @@ namespace ncore
         }
     }
 
-    void g_qsort(void* a, s32 n, s32 es, s32 (*cmp)(const void*, const void*, void*), void* user_data)
+    void g_qsort(void* a, s32 n, s32 es, s8 (*cmp)(const void*, const void*, void*), void* user_data)
 	{
 		switch (es)
 		{
